refactor: refactored code
