+++ src/Helper/commonFunction.js
