+++ src/Helper/index.js
@@ -1,5 +1,6 @@
-import { getFilteredBySearchData, getFilteredData } from "./filterFunciton";
-import { findItemInArray } from "./commonFuction";
+import { getFilteredBySearchData, getFilteredData } from "./filterFunction";
+import { findItemInArray } from "./commonFunction";
 import {watchLaterHandler, likedVideoHandler, removeVideoFromHistory, removeFromLikedVideo, addToLikedVideos, calculateView, historyHandler} from "./serviceFunctions";
+import { newPlaylistHandler } from "./playlistFunction";
 
-export {getFilteredBySearchData, getFilteredData, findItemInArray, watchLaterHandler, likedVideoHandler, removeVideoFromHistory, removeFromLikedVideo, addToLikedVideos, calculateView, historyHandler}
+export {getFilteredBySearchData, getFilteredData, findItemInArray, watchLaterHandler, likedVideoHandler, removeVideoFromHistory, removeFromLikedVideo, addToLikedVideos, calculateView, historyHandler, newPlaylistHandler}
