+++ src/pages/Home/Home.jsx
@@ -1,6 +1,7 @@
 import { Navbar, Sidebar, VideoCard } from "../../Components/index"
 import { useFilter } from "../../Context";
-import { TabButtonChips } from "./tabsButtonChips";
+import { PlaylistModal } from "./PlaylistModal/PlaylistModal";
+import { TabButtonChips } from "./TabsButtonChips";
 
 export const Home = () => {
     const {filteredVideoList} = useFilter();
@@ -8,6 +9,7 @@ export const Home = () => {
         <div className="body-template">
             <Navbar />
             <Sidebar />
+            <PlaylistModal />
             <main className="main-box flex-column gap-2">
                 <TabButtonChips />
                 <div className="videoCard-box flex-row align-center justify-evenly gap-2">
