+++ src/pages/index/index.css
@@ -3,3 +3,4 @@
 @import "../UserProfile/userProfile.css";
 @import "../SingleVideoPage/singleVideo.css";
 @import "../HistoryPage/history.css";
+@import "../Home/PlaylistModal/playlistModal.css";
